cleanup code after review
