+++ userdocs/index.rst
@@ -18,7 +18,7 @@ Welcome to the Warewulf User Guide!
    Warewulf Initialization <contents/initialization>
    Container Management <contents/containers>
    Kernel Management <contents/kernel>
-   Boot Management <contents/grub>
+   Boot Management <contents/boot-management>
    Node Configuration <contents/nodeconfig>
    Node Profiles <contents/profiles>
    Warewulf Overlays <contents/overlays>
